requirements.txt
aiodns==2.0.0
aiohttp==2.3.10
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
circuitbreaker==2.0.0
click==8.0.4
click-plugins==1.1.1
colorama==0.4.6
config==0.5.1
cryptography==43.0.3
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.2
dnspython==1.15.0
dnspython3==1.15.0
filelock==3.16.1
fonttools==4.54.1
future==1.0.0
geocoder==1.38.1
idna==2.9
idna-ssl==1.1.0
ipwhois==1.3.0
Jinja2==3.1.4
jmespath==0.10.0
lxml==5.3.0
MarkupSafe==3.0.1
multidict==6.1.0
oci==2.137.0
oci-cli==3.49.3
pdfkit==1.0.0
pillow==10.4.0
ping3==4.0.8
prompt-toolkit==3.0.43
propcache==0.2.0
pycares==4.4.0
pycparser==2.22
pydnsbl==1.1.6
pydyf==0.11.0
pyOpenSSL==24.2.1
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-whois==0.9.4
pytz==2024.2
PyYAML==6.0.1
ratelim==0.1.6
requests==2.32.3
requests-file==2.1.0
scapy==2.6.0
shodan==1.31.0
six==1.16.0
soupsieve==2.6
spam-lists==1.0.0
sqlparse==0.5.1
terminaltables==3.1.10
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.1.2
tqdm==4.23.4
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
XlsxWriter==3.2.0
yarl==1.15.5
zopfli==0.2.3.post1
